requirements.txt
numpy==1.23.1
scikit-learn==1.1.1
matplotlib==3.5.1
torch==1.11.0
flask==2.1.1
pillow==9.2.0
gunicorn==20.1.0
